+++ app/configurator/components/chart-type-selector.tsx
@@ -17,6 +17,7 @@ import { ControlSectionSkeleton } from "@/configurator/components/chart-controls
 import { IconButton } from "@/configurator/components/icon-button";
 import { useAddOrEditChartType } from "@/configurator/config-form";
 import { ConfiguratorStateWithChartConfigs } from "@/configurator/configurator-state";
+import { useFlag } from "@/flags";
 import { useDataCubesComponentsQuery } from "@/graphql/hooks";
 import { useLocale } from "@/locales/use-locale";
 
@@ -39,6 +40,7 @@ export const ChartTypeSelector = (props: ChartTypeSelectorProps) => {
     chartKey,
     ...rest
   } = props;
+  const enableBarChart = useFlag("enable-experimental-features");
   const locale = useLocale();
   const chartConfig = getChartConfig(state);
   const [{ data }] = useDataCubesComponentsQuery({
@@ -121,7 +123,11 @@ export const ChartTypeSelector = (props: ChartTypeSelectorProps) => {
                 message: "Regular",
               })}
               currentChartType={chartType}
-              chartTypes={regularChartTypes}
+              chartTypes={
+                enableBarChart
+                  ? regularChartTypes
+                  : regularChartTypes.filter((chartType) => chartType !== "bar")
+              }
               possibleChartTypesDict={possibleChartTypesDict}
               onClick={handleClick}
               testId="chart-type-selector-regular"